+++ cmd/downloader/main.go
@@ -96,21 +96,14 @@ func run(ctx context.Context) error {
 	return srv.Shutdown(sctx)
 }
 
-func getMaxConnectionAge(maxConnectionAge int) int {
-	if maxConnectionAge < 60 {
-		maxConnectionAge = 15 * 60
-	}
-	return maxConnectionAge
-}
-
 func newGrpcServer(svc geogrpc.GeocubeDownloaderService, maxConnectionAgeValue int, chunkSizeBytes int) *grpc.Server {
 	opts := []grpc.ServerOption{
 		grpc.KeepaliveParams(keepalive.ServerParameters{
-			MaxConnectionAge:      time.Duration(getMaxConnectionAge(maxConnectionAgeValue)) * time.Second,
+			MaxConnectionAge:      time.Duration(maxConnectionAgeValue) * time.Second,
 			MaxConnectionAgeGrace: time.Minute})}
 
 	grpcServer := grpc.NewServer(opts...)
-	pb.RegisterGeocubeDownloaderServer(grpcServer, geogrpc.NewDownloader(svc, getMaxConnectionAge(maxConnectionAgeValue), chunkSizeBytes))
+	pb.RegisterGeocubeDownloaderServer(grpcServer, geogrpc.NewDownloader(svc, maxConnectionAgeValue, chunkSizeBytes))
 	return grpcServer
 }
 
@@ -119,7 +112,7 @@ func newDownloaderAppConfig() (*serverConfig, error) {
 
 	flag.BoolVar(&serverConfig.TLS, "tls", false, "enable TLS protocol")
 	flag.StringVar(&serverConfig.AppPort, "port", "8080", "geocube downloader port to use")
-	flag.IntVar(&serverConfig.MaxConnectionAge, "maxConnectionAge", 0, "grpc max age connection")
+	flag.IntVar(&serverConfig.MaxConnectionAge, "maxConnectionAge", 15*60, "grpc max age connection in seconds")
 	flag.IntVar(&serverConfig.CubeWorkers, "workers", 1, "number of workers to parallelize the processing of the slices of a cube (see also GdalMultithreading)")
 	flag.IntVar(&serverConfig.ChunkSizeByte, "chunk-size", 1024*1024, "chunk size for grpc streaming of images in bytes. If an image is bigger than chunk_size_bytes, it is divided into chunks and streamed. Grpc recommends a chunk_size of 64kbytes, but in localhost, performances are better with a bigger chunk_size, such as 1Mbytes. By default, chunk_size is limited by Grpc to 4Mbytes.")
 	serverConfig.GDALConfig = cmd.GDALConfigFlags()